Ripped out unused analyzer functions from earlier attempt at projection reordering for subqueries.
