+++ sql/analyzer/resolve_columns.go
@@ -539,47 +539,6 @@ func resolveColumnExpression(ctx *sql.Context, a *Analyzer, e column, columns ma
 	), nil
 }
 
-// resolveGetIndexedFields attempts to resolve GetIndexedField expressions in the plan. Such fields are placeholders to
-// handle projections etc. on expressions that can't be resolved yet, and can be replaced by normal GetField expressions
-// once their children are resolved.
-func resolveGetIndexedFields(ctx *sql.Context, a *Analyzer, n sql.Node, scope *Scope) (sql.Node, error) {
-	return plan.TransformExpressionsUpWithNode(n, func(n sql.Node, e sql.Expression) (sql.Expression, error) {
-		gif, ok := e.(*expression.GetIndexedField)
-		if !ok {
-			return e, nil
-		}
-
-		// If the expression in the child node that this expression refers to is resolved, we can replace it with a normal
-		// GetField expression. We could just check the child expression itself, but just wait for the entire child to be
-		// resolved. It makes it easier to get the schema of the field being resolved.
-		if !allResolved(n.Children()) {
-			return e, nil
-		}
-
-		childSchema := schema(n.Children())
-
-		return expression.NewGetField(
-			gif.Index(), childSchema[gif.Index()].Type, gif.Name(), childSchema[gif.Index()].Nullable), nil
-	})
-}
-
-func allResolved(nodes []sql.Node) bool {
-	for _, n := range nodes {
-		if !n.Resolved() {
-			return false
-		}
-	}
-	return true
-}
-
-func schema(nodes []sql.Node) sql.Schema {
-	var schema sql.Schema
-	for _, n := range nodes {
-		schema = append(schema, n.Schema()...)
-	}
-	return schema
-}
-
 // pushdownGroupByAliases reorders the aggregation in a groupby so aliases defined in it can be resolved in the grouping
 // of the groupby. To do so, all aliases are pushed down to a projection node under the group by.
 func pushdownGroupByAliases(ctx *sql.Context, a *Analyzer, n sql.Node, scope *Scope) (sql.Node, error) {
