+++ sql/analyzer/resolve_subqueries.go
@@ -1,9 +1,7 @@
 package analyzer
 
 import (
-	"fmt"
 	"github.com/liquidata-inc/go-mysql-server/sql"
-	"github.com/liquidata-inc/go-mysql-server/sql/expression"
 	"github.com/liquidata-inc/go-mysql-server/sql/plan"
 )
 
@@ -58,115 +56,4 @@ func resolveSubqueryExpressions(ctx *sql.Context, a *Analyzer, n sql.Node, scope
 
 		return s.WithQuery(analyzed), nil
 	})
-}
-
-// pullUpMissingSubqueryColumns examines subqueries to see which columns from outer scopes are missing in the scope
-// node, and pulls them up. An additional higher-level projection is added as necessary.
-func pullUpMissingSubqueryColumns(ctx *sql.Context, a *Analyzer, n sql.Node, scope *Scope) (sql.Node, error) {
-	return plan.TransformUp(n, func(n sql.Node) (sql.Node, error) {
-		e, ok := n.(sql.Expressioner)
-		if !ok {
-			return n, nil
-		}
-
-		var deferredColumns []*deferredColumn
-		var exprs []sql.Expression
-		for _, e := range e.Expressions() {
-			// var subquery sql.Expression
-			// if a, ok := e.(*expression.Alias); ok {
-			//
-			// }
-			// TODO: handle aliases
-			s, ok := e.(*plan.Subquery)
-			if !ok {
-				exprs = append(exprs, e)
-				continue
-			}
-
-			// wrap any unaliased subqueries in an alias so that they can be identified by parent nodes during further analysis
-			exprs = append(exprs, expression.NewAlias(s.QueryString, s))
-			deferredColumns = append(deferredColumns, findDeferredColumns(s.Query)...)
-		}
-
-		if len(deferredColumns) > 0 {
-			switch n.(type) {
-			case *plan.Project:
-				var err error
-				n, err = replaceExpressions(n, exprs)
-				if err != nil {
-					return nil, err
-				}
-			}
-
-			return addDeferredColumns(n, deferredColumns)
-		}
-
-		return n, nil
-	})
-}
-
-func replaceExpressions(n sql.Node, exprs []sql.Expression) (sql.Node, error) {
-	switch nn := n.(type) {
-	case sql.Expressioner:
-		return nn.WithExpressions(exprs...)
-	default:
-		panic(fmt.Sprintf("Cannot replace expressions for node %T", n))
-	}
-}
-
-// addDeferredColumns adds the given deferred columns to necessary nodes in the tree given, as well as a top-level
-// projection to restore the original schema.
-func addDeferredColumns(n sql.Node, columns []*deferredColumn) (sql.Node, error) {
-	return plan.TransformUp(n, func(n sql.Node) (sql.Node, error) {
-		switch n := n.(type) {
-		case *plan.Project:
-			// TODO: error check to make sure that these expressions aren't already included here
-			nn := plan.NewProject(append(n.Expressions(), deferredColumnsToUnresolvedColumns(columns)...), n.Child)
-			return plan.NewProject(projectionsToGetFields(n.Projections), nn), nil
-		default:
-			return n, nil
-		}
-	})
-}
-
-func projectionsToGetFields(projections []sql.Expression) []sql.Expression {
-	getFields := make([]sql.Expression, len(projections))
-	for i, projection := range projections {
-		if projection.Resolved() {
-			getFields[i] = expression.NewGetField(i, projection.Type(), getName(projection), projection.IsNullable())
-		} else {
-			getFields[i] = expression.NewGetIndexedField(i, getName(projection))
-		}
-	}
-	return getFields
-}
-
-func getName(e sql.Expression) string {
-	if n, ok := e.(sql.Nameable); ok {
-		return n.Name()
-	}
-	if _, ok := e.(*plan.Subquery); ok {
-		return "subquery"
-	}
-	return e.String()
-}
-
-func deferredColumnsToUnresolvedColumns(dcs []*deferredColumn) []sql.Expression {
-	ucs := make([]sql.Expression, len(dcs))
-	for i, dc := range dcs {
-		ucs[i] = expression.NewUnresolvedQualifiedColumn(dc.Table(), dc.Name())
-	}
-	return ucs
-}
-
-func findDeferredColumns(n sql.Node) []*deferredColumn {
-	var cols []*deferredColumn
-	plan.InspectExpressions(n, func(e sql.Expression) bool {
-		if dc, ok := e.(*deferredColumn); ok {
-			cols = append(cols, dc)
-		}
-		return true
-	})
-
-	return cols
 }
