+++ sql/expression/get_field.go
@@ -149,73 +149,3 @@ func (f *GetSessionField) WithChildren(children ...sql.Expression) (sql.Expressi
 	}
 	return f, nil
 }
-
-// GetIndexedField is an expression to get the field of a table.
-type GetIndexedField struct {
-	fieldIndex int
-	name       string
-}
-
-// NewGetIndexedField creates a new GetIndexedField expression.
-func NewGetIndexedField(index int, fieldName string) *GetIndexedField {
-	return &GetIndexedField{
-		fieldIndex: index,
-		name:       fieldName,
-	}
-}
-
-var _ sql.Expression = (*GetIndexedField)(nil)
-
-// Index returns the index where the GetField will look for the value from a sql.Row.
-func (p *GetIndexedField) Index() int { return p.fieldIndex }
-
-// Children implements the Expression interface.
-func (*GetIndexedField) Children() []sql.Expression {
-	return nil
-}
-
-// Resolved implements the Expression interface.
-func (p *GetIndexedField) Resolved() bool {
-	return false
-}
-
-// Name implements the Nameable interface.
-func (p *GetIndexedField) Name() string { return p.name }
-
-// IsNullable returns whether the field is nullable or not.
-func (p *GetIndexedField) IsNullable() bool {
-	panic("GetIndexedField is a placeholder expression, but IsNullable was called")
-}
-
-// Type returns the type of the field.
-func (p *GetIndexedField) Type() sql.Type {
-	panic("GetIndexedField is a placeholder expression, but Type was called")
-}
-
-// Eval implements the Expression interface.
-func (p *GetIndexedField) Eval(ctx *sql.Context, row sql.Row) (interface{}, error) {
-	panic("GetIndexedField is a placeholder expression, but Eval was called")
-}
-
-// WithChildren implements the Expression interface.
-func (p *GetIndexedField) WithChildren(children ...sql.Expression) (sql.Expression, error) {
-	if len(children) != 0 {
-		return nil, sql.ErrInvalidChildrenNumber.New(p, len(children), 0)
-	}
-	return p, nil
-}
-
-func (p *GetIndexedField) String() string {
-	return fmt.Sprintf("%s", p.name)
-}
-
-func (p *GetIndexedField) DebugString() string {
-	return fmt.Sprintf("[%s, idx=%d]", p.name, p.fieldIndex)
-}
-
-// WithIndex returns this same GetField with a new index.
-func (p *GetIndexedField) WithIndex(n int) sql.Expression {
-	p2 := *p
-	p2.fieldIndex = n
-	return &p2
-}
